microsoftgraph/msgraph-sdk-java#211 fix ChunkedUpload limit

+++ src/main/java/com/microsoft/graph/concurrency/ChunkedUploadProvider.java
@@ -82,7 +82,7 @@ public class ChunkedUploadProvider<UploadType> {
     /**
      * The stream size
      */
-    private final int streamSize;
+    private final long streamSize;
 
     /**
      * The upload response handler
@@ -106,7 +106,7 @@ public class ChunkedUploadProvider<UploadType> {
     public ChunkedUploadProvider(final UploadSession uploadSession,
                                  final IGraphServiceClient client,
                                  final InputStream inputStream,
-                                 final int streamSize,
+                                 final long streamSize,
                                  final Class<UploadType> uploadTypeClass) {
         if (uploadSession == null) {
             throw new InvalidParameterException("Upload session is null.");
